# openapi.yaml
openapi: 3.1.0
info:
  title: FindTutor API
  version: 0.1.0
  description: >
    Contract for the FindTutor language-tutoring marketplace (Uzbekistan).
    Hand-written and treated as the source of truth: the Go backend renders
    responses to match it, and the Next.js frontend generates its typed client
    from it (openapi-typescript + openapi-fetch).

    Money is always an integer amount in MINOR units (tiyin for UZS, cents for
    USD) plus a currency code — never a float.

    Language: every human-readable `error.message` is rendered in the language
    of the request's `Accept-Language` header (`en`, `ru` or `uz`; anything
    else falls back to English) and the response carries `Content-Language`.
    Email is written in the account's stored `locale` (see AuthUser), set from
    Accept-Language at registration and changed via PATCH /v1/auth/me.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/bookings:
    post:
      tags: [bookings]
      summary: Book a lesson
      description: >
        Creates a booking in `pending_payment` and opens a `requires_payment`
        payment intent for its price. The server re-derives the bookable slot
        set and rejects a `start_at` that is not a real, currently free slot; it
        never trusts a client-supplied price.

        `duration_minutes` must be one of 30/60/90/120. When `is_trial` is true
        the duration is forced to 30 minutes and the price is the teacher's
        `trial_price_minor` (400 if the teacher offers no trial). You cannot
        book your own teacher profile.

        The student then calls `POST /v1/bookings/{id}/pay` to authorise
        payment, which confirms the booking.
      operationId: createBooking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBookingRequest" }
      responses:
        "201":
          description: The created booking.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: >
            The slot is not bookable (`slot_unavailable` — outside availability,
            misaligned, in the past, or already taken) or was taken by a
            concurrent request between the check and the insert (`slot_taken`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [bookings]
      summary: List the caller's bookings
      description: >
        With no `role`, returns every booking the caller participates in (as
        student or as the owner of the teacher profile). `role=student` limits
        it to bookings the caller booked; `role=teacher` to bookings for the
        teacher profile the caller owns (empty if they own none). Newest lesson
        first. Optional `status` filter.
      operationId: listBookings
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          required: false
          schema: { type: string, enum: [student, teacher] }
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/BookingStatus" }
      responses:
        "200":
          description: The matching bookings.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookingList" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/bookings/{id}:
    get:
      tags: [bookings]
      summary: Get one booking
      operationId: getBooking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The booking.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/bookings/{id}/resources:
    post:
      tags: [resources]
      summary: Attach a resource to a lesson
      description: >
        Teacher-owner only. The resource must belong to the same teacher, be
        `published`, and not archived. `kind: homework` is rejected for
        `material` / `article` resources (400 — they have no submission flow;
        attach them as `material` instead). Attaching the same resource to the
        same booking twice is a 409 (`resource_already_attached`), not a
        duplicate row.
      operationId: attachBookingResource
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AttachResourceRequest" }
      responses:
        "201":
          description: The created attachment.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttachedResource" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The caller is not the teacher for this lesson, or does not own the resource (`forbidden`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: "`resource_already_attached`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [resources]
      summary: A lesson's attached resources
      description: >
        Participant-only. The teacher sees the full resource content
        (including quiz answers) and no submission; the student sees every
        question's `correct` field stripped and their own submission summary
        per attachment.
      operationId: listBookingResources
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The lesson's attachments.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttachedResourceList" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/bookings/{id}/resources/{attachmentId}:
    delete:
      tags: [resources]
      summary: Detach a resource from a lesson
      description: >
        Teacher-owner only. Any submissions already filed against the
        resource are kept — detaching does not delete history.
      operationId: detachBookingResource
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: attachmentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Detached. }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/submissions:
    post:
      tags: [resources]
      summary: Start (or resume) a submission
      description: >
        Body must carry exactly one of `booking_id` (lesson-context homework)
        or `enrollment_id` (phase C2, a course-embedded resource). Idempotent:
        a second call for the same resource/booking (or resource/enrollment)
        returns the existing submission rather than erroring.

        Lesson context: the resource must be attached to the booking as
        `homework`, and the caller must be the booking's student (409
        `homework_not_assigned` / 403 otherwise).

        Course context: the resource must actually be a curriculum item of
        the enrollment's course (409 `resource_not_in_course` otherwise), and
        the caller must be the enrollment's student.

        Either way the resource type must carry a submission flow (`quiz` /
        `listening` / `reading` / `writing`).
      operationId: startSubmission
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StartSubmissionRequest" }
      responses:
        "201":
          description: The submission (new or already in progress).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Submission" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The caller is not the booking's/enrollment's student (`forbidden`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: "No booking/enrollment with that id (`not_found`)."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: >
            `homework_not_assigned` — the resource isn't attached to this
            booking as homework — or `resource_not_in_course` — the resource
            isn't part of the enrolled course.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [resources]
      summary: The teacher's grading inbox
      description: >
        The caller's own resources' submissions. `status` defaults to
        `submitted` (i.e. `writing` submissions awaiting a grade); `all` lifts
        the filter. Newest-submitted-first.
      operationId: listSubmissionInbox
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [in_progress, submitted, graded, all], default: submitted }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200":
          description: A page of the inbox.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubmissionList" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: "`no_teacher_profile`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/submissions/{id}:
    patch:
      tags: [resources]
      summary: Save draft answers
      description: Owner-only, and only while the submission is `in_progress`.
      operationId: saveSubmissionAnswers
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SaveAnswersRequest" }
      responses:
        "200":
          description: The updated submission.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Submission" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: "`invalid_state` — the submission is no longer `in_progress`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [resources]
      summary: Get one submission
      description: The owning student or the booking's teacher-owner only.
      operationId: getSubmission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The submission.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Submission" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/submissions/{id}/submit:
    post:
      tags: [resources]
      summary: Submit a homework
      description: >
        Owner-only, only from `in_progress`. `quiz` / `listening` / `reading`
        auto-grade against the resource's questions and land directly on
        `graded` (`auto_score` / `auto_max` set, all-or-nothing per question,
        no partial credit). `writing` lands on `submitted` and awaits a
        teacher's grade.
      operationId: submitSubmission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The finalized submission.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Submission" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: "`invalid_state` — the submission is not `in_progress`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/submissions/{id}/grade:
    post:
      tags: [resources]
      summary: Grade a writing submission
      description: >
        Only a `writing` submission, only from `submitted`. Caller must be the
        teacher-owner of the submission's booking. `score` is nullable — a
        teacher may grade with feedback only. Best-effort emails the student;
        a mail failure never fails the request.
      operationId: gradeSubmission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GradeSubmissionRequest" }
      responses:
        "200":
          description: The graded submission.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Submission" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The caller is not the teacher-owner of this submission's booking (`forbidden`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: "`invalid_state` — the submission is not `submitted`, or is not a writing task."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /v1/resources:
    get:
      tags: [resources]
      summary: The caller's resource library
      operationId: listResources
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: type, in: query, schema: { $ref: "#/components/schemas/ResourceType" } }
        - { name: status, in: query, schema: { type: string, enum: [draft, published] } }
        - { name: archived, in: query, schema: { type: string, enum: ["true"] }, description: Include archived resources. }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200":
          description: A page of the library.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResourceList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: "`no_teacher_profile` — the caller has no teacher profile."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    post:
      tags: [resources]
      summary: Create a resource
      operationId: createResource
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateResourceRequest" }
      responses:
        "201":
          description: The created resource.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Resource" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: "`no_teacher_profile`."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  /v1/resources/{id}/publish:
    post:
      tags: [resources]
      summary: Publish a resource
      operationId: publishResource
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: The resource, now published., content: { application/json: { schema: { $ref: "#/components/schemas/Resource" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/resources/{id}/unpublish:
    post:
      tags: [resources]
      summary: Return a resource to draft
      operationId: unpublishResource
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: The resource, now a draft., content: { application/json: { schema: { $ref: "#/components/schemas/Resource" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived HS256 access token from /v1/auth/{login,register,refresh}.
    cookieAuth:
      type: apiKey
      in: cookie
      name: ftr_session
      description: >
        HttpOnly opaque refresh token, Path=/v1/auth, SameSite=Lax. Set by the
        auth endpoints; the browser sends it automatically to /v1/auth/refresh
        and /v1/auth/logout. Not readable or settable by client JS.
  responses:
    BadRequest:
      description: The request was malformed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    CreateBookingRequest:
      type: object
      required: [teacher_slug, start_at, duration_minutes]
      properties:
        teacher_slug: { type: string }
        start_at:
          type: string
          format: date-time
          description: RFC3339 UTC. Must be in the future and a real bookable slot.
        duration_minutes:
          type: integer
          enum: [30, 60, 90, 120]
          description: Ignored when is_trial is true (forced to 30).
        is_trial:
          type: boolean
          default: false
          description: A trial lesson — 30 minutes, priced at the teacher's trial_price.
...